Update ActivitySubscriptionEmailHelper.java
The way the weekly report is sent is totally modified.
Now it is not necessary to be subscribed to the activities to receive the weekly report, it is only necessary to enable the sending of weekly notifications in the user's profile.

Retrieve the Challenges that have been created after a certain date.

Allows to distinguish between Conversation Threads and Prototyping Spaces

The message sent to the user is personalized.